requirements.txt
python-bitcoinrpc